requirements.txt
aiohttp==3.11.16
fastapi==0.115.12
gunicorn==23.0.0
librosa==0.11.0
psutil==7.0.0
pydantic==2.11.2
pydub==0.25.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
torch==2.6.0
torchaudio==2.6.0
coqui-tts==0.26.0
coqui-tts-trainer==0.2.3
uvicorn==0.34.0
uvloop==0.21.0
python-multipart==0.0.20
coqpit-config==0.2.1
